Make the keybindings module compile on mobile
The keybindings work added editable-focus handlers (editable_focus_in/out,
EditableFocused) to inputs across the UI, but the whole module was gated
#[cfg(feature = "desktop")] because its shortcut engine depends on muda
(desktop menus). The mobile/iOS build then failed to resolve those
handlers at ~30 call sites.

Split the module: the platform-neutral focus primitives move to an always-
compiled keybindings/mod.rs; the muda-dependent shortcut engine stays in a
desktop-only keybindings/desktop.rs, re-exported on desktop. No behavior
change on desktop; unblocks the iOS check.
