+++ src/ui/mod.rs
@@ -4,7 +4,6 @@ pub mod components;
 pub mod graph_view;
 pub mod helpers;
 pub mod import_export;
-#[cfg(feature = "desktop")]
 pub mod keybindings;
 pub mod layout;
 pub mod library;
